Move TargetLine and TargetParams definitions into shared module
They are only use by the outer layer, and do not need to be exported.

Decide that equality should be reflexive on TargetLines
Using the table as a proxy for equality of devices is not done w/ == so
we can view equality on TargetLines simply.
