+++ src/lineardev.rs
@@ -11,9 +11,9 @@ use super::deviceinfo::DeviceInfo;
 use super::dm::{DM, DmFlags};
 use super::result::{DmResult, DmError, ErrorEnum};
 use super::segment::Segment;
-use super::shared::{DmDevice, device_create, device_exists, device_match, parse_device,
-                    table_reload};
-use super::types::{DevId, DmName, DmUuid, Sectors, TargetLine, TargetParams, TargetTypeBuf};
+use super::shared::{DmDevice, TargetLine, TargetParams, device_create, device_exists,
+                    device_match, parse_device, table_reload};
+use super::types::{DevId, DmName, DmUuid, Sectors, TargetTypeBuf};
 
 
 #[derive(Debug, Eq, PartialEq)]
