+++ src/thinpooldev.rs
@@ -13,10 +13,9 @@ use super::dm::{DM, DmFlags};
 use super::lineardev::LinearDev;
 use super::result::{DmResult, DmError, ErrorEnum};
 use super::segment::Segment;
-use super::shared::{DmDevice, device_create, device_exists, device_match, parse_device,
-                    table_reload};
-use super::types::{DataBlocks, DevId, DmName, DmUuid, MetaBlocks, Sectors, TargetLine,
-                   TargetParams, TargetTypeBuf};
+use super::shared::{DmDevice, TargetLine, TargetParams, device_create, device_exists,
+                    device_match, parse_device, table_reload};
+use super::types::{DataBlocks, DevId, DmName, DmUuid, MetaBlocks, Sectors, TargetTypeBuf};
 
 #[cfg(test)]
 use std::path::Path;
