+++ src/types.rs
@@ -21,7 +21,6 @@ use std::fmt;
 use std::iter::Sum;
 use std::mem::transmute;
 use std::ops::{Deref, Div, Mul, Rem, Add};
-use std::str::FromStr;
 
 use serde;
 
@@ -423,23 +422,6 @@ const DM_TARGET_TYPE_LEN: usize = 16;
 
 str_id!(TargetType, TargetTypeBuf, DM_TARGET_TYPE_LEN, str_check);
 
-/// The trait for properties of the params string of TargetType
-pub trait TargetParams: fmt::Debug + fmt::Display + Eq + FromStr + PartialEq {}
-
-impl TargetParams for String {}
-
-/// One line of a device mapper table.
-#[derive(Debug, PartialEq)]
-pub struct TargetLine<T: TargetParams> {
-    /// The start of the segment
-    pub start: Sectors,
-    /// The length of the segment
-    pub length: Sectors,
-    /// The target type
-    pub target_type: TargetTypeBuf,
-    /// The target specific parameters
-    pub params: T,
-}
 
 #[cfg(test)]
 mod tests {
